Directory refactor (placing the header-related components in their own directory).
